Refactors the bottom navigation bar into two reusable actions and removes the old bottom navigation bar tests and component file
